feat(repository): add RideRepository as single source of truth
- Expose friend, ride and participant operations to ViewModels
- addRide inserts ride then links participants using generated id
- deleteRide relies on CASCADE to clean up RideParticipant rows
- calculateOwedAmounts computes each friend's monthly share via
  totalFare / participantCount summed across all their rides

+++ app/src/main/java/com/example/farespliter/data/dao/FriendDao.kt
@@ -13,7 +13,7 @@ interface FriendDao{
     suspend fun insert(friend: Friend)
 
     @Query("DELETE FROM friends WHERE id = :friendId")
-    suspend fun deleteById(friendId: Int)
+    suspend fun deleteById(friendId: Long)
 
     @Query("SELECT * FROM friends ORDER BY name ASC")
     fun getAllFriends() : LiveData<List<Friend>>
